fix: improve `unpartial(base, partial)` type
type for `unpartial(parent, base, partial)` is marked as deprecated.
The type is loosen now similar to `required(parent, base, partial)` to make it more flexible.
But note that it is incorrect as the values are not deeply merged.
